requirements.txt
google-cloud-storage
PyPDF2
